requirements.txt
aiodns==1.1.1
aiohttp==3.4.4
alembic==1.0.0
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
cchardet==2.1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
future==0.16.0
idna==2.7
idna-ssl==1.1.0
Mako==1.0.7
MarkupSafe==1.0
multidict==4.4.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycares==2.3.0
pycparser==2.19
PySocks==1.6.8
python-dateutil==2.7.3
python-editor==1.0.3
python-telegram-bot==11.1.0
six==1.11.0
SQLAlchemy==1.2.12
yarl==1.2.6
requests==2.20.0
